requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-extensions==2.0.7
django-filter==1.1.0
django-postgres-extensions==0.9.3
django-taggit==0.22.2
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
lxml==4.2.4
psycopg2==2.7.4
pytz==2018.5
six==1.11.0
uWSGI==2.0.17
Markdown==2.6.11
